Creades dataclass per a missatgeria
